web-search-agent/pyproject.toml
[project]
name = "web-search-agent"
version = "0.1.0"
description = "A ReAct web search agent powered by Gemini API"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "ddgs>=9.0.0",
    "python-dotenv>=1.0.0",
]
